+++ LeaveSystemMVC/Models/sleaveBalanceModel.cs
@@ -11,24 +11,24 @@ namespace LeaveSystemMVC.Models
 
         public int empId { get; set; } = 0;
         [Required]
-        public int annual { get; set; } = 0;
+        public decimal annual { get; set; } = 0;
         public int annualID { get; set; } = 0;
         [Required]
-        public int sick { get; set; } = 0;
+        public decimal sick { get; set; } = 0;
         public int sickID { get; set; } = 0;
         [Required]
-        public int compassionate { get; set; } = 0;
+        public decimal compassionate { get; set; } = 0;
         public int compassionateID { get; set; } = 0;
         [Required]
-        public int maternity { get; set; } = 0;
+        public decimal maternity { get; set; } = 0;
         public int maternityID { get; set; } = 0;
         [Required]
-        public int daysInLieue { get; set; } = 0;
+        public decimal daysInLieue { get; set; } = 0;
         public int daysInLieueID { get; set; } = 0;
-        public int unpaidTotal { get; set; } = 0;
+        public decimal unpaidTotal { get; set; } = 0;
         public int unpaidID { get; set; } = 0;
         [Required]
-        public int shortLeaveHours { get; set; } = 0;
+        public decimal shortLeaveHours { get; set; } = 0;
         public int shortID { get; set; } = 0;
     }
